fix:修复nacos java-agent双注册双发现迁移

+++ polaris-agent-build/bin/build.sh
@@ -28,7 +28,12 @@ mkdir -p "${folder_name}/plugins"
 cp "polaris-agent-build/conf/polaris-agent.config" "${folder_name}/conf"
 
 echo "start to build package"
-docker run --rm -u root -v "$(pwd)":/home/maven/project -w /home/maven/project maven:3.8.6-openjdk-8 mvn clean -B package --file pom.xml
+
+if [[ "${use_docker_env}" == "true" ]]; then
+  docker run --rm -u root -v "$(pwd)":/home/maven/project -w /home/maven/project maven:3.8.6-openjdk-8 mvn clean -B package --file pom.xml
+else
+  mvn clean -B package --file pom.xml
+fi
 
 cp "polaris-agent-core/polaris-agent-core-bootstrap/target/polaris-agent-core-bootstrap.jar" "${folder_name}/"
